# docs/openapi/marketflow-sprint-01-02.openapi.yaml
openapi: 3.0.3
info:
  title: MarketFlow Sprint 1 and 2 APIs
  version: 0.2.0
  description: |
    Contract for the Identity, Catalog and Order services. Browser clients use
    the gateway prefixes /api/identity, /api/catalog and /api/order.
servers:
  - url: http://localhost:8080/api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
  /identity/auth/register:
    post:
      summary: Register a Customer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, displayName, password]
              properties:
                email: { type: string, format: email }
                displayName: { type: string }
                password: { type: string, format: password, minLength: 8 }
      responses:
        '201': { description: Customer account and access token created }
        '400': { description: Invalid request }
        '409': { description: Email already exists }
  /identity/auth/login:
    post:
      summary: Sign in
      responses:
        '200': { description: Access token and user roles returned }
        '401': { description: Invalid credentials }
  /identity/admin/users:
    get:
      summary: List accounts
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Operations Admin account list }
        '403': { description: Operations Admin role required }
    post:
      summary: Create staff or courier account
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Account created }
        '403': { description: Operations Admin role required }
...